requirements.txt
Django==4.1
Pillow==9.2.0
autopep8==1.7.0
environs==9.5.0
django-crispy-forms==1.14.0
django-ckeditor-6.5.0